refactor(workflows): neutralize definition and launch service naming
