+++ turbo/apps/api/src/signals/services/workflow-automation-access.service.ts
@@ -6,7 +6,7 @@ import type { ReadonlyDb } from "../external/db";
 import {
   loadVisibleWorkflowById,
   type WorkflowAgentInfo,
-} from "./zero-workflow-data.service";
+} from "./workflow-data.service";
 
 type WorkflowAutomationRow = typeof workflowAutomations.$inferSelect;
