+++ turbo/apps/api/src/signals/services/workflow-automation.service.ts
@@ -80,7 +80,7 @@ import {
   visibleWorkflowCondition,
   workflowSummary,
   type WorkflowMember,
-} from "./zero-workflow-data.service";
+} from "./workflow-data.service";
 import {
   ensureGmailWatchForUser,
   hasEnabledGmailConsumer,
@@ -123,7 +123,7 @@ import {
 } from "./workflow-webhook-automation.service";
 import { dispatchFailedRunCallbacks } from "./agent-run-callback.service";
 import { runWorkflowAutomationNow$ } from "./workflow-automation-run.service";
-import type { RunWorkflowAutomationResult } from "./zero-workflow-automation-launch.service";
+import type { RunWorkflowAutomationResult } from "./workflow-automation-launch.service";
 import { manualTriggerSource } from "./workflow-automation-trigger-source";
 import {
   ensureWorkflowUserAutomationThread,
